+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `COINGECKO` as a new data source type
 - Added the configuration to publish a `linux/arm/v7` docker image
 
 ### Fixed